IMPALA-14312: Fix Issues with ssl_minimum_version Flag Validator
The validator for the ssl_minimum_version flag has several issues
that are fixed.

Allows flag to be empty as long as both internal and external TLS is
not configured.

Fixes allowed value for TLS v1 to be tlsv1 instead of the incorrect
value tlsv1.0.

Removes "tlsv1.3" as an allowed value since Thrift does not support
that value as the minimum TLS version.

Testing accomplished by new ctests and manual testing.

Change-Id: I6493852b581e26c203b6b46b97be76100bcc534b
Reviewed-on: http://gerrit.cloudera.org:8080/23300
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Jason Fehr <jfehr@cloudera.com>