IMPALA-14101: [part 1] Commit Cost file from Calcite
This commit is just a copy of the VolcanoCost.java file from Calcite
into this Impala repository.  The file can be found here:

https://github.com/apache/calcite/blob/calcite-1.37.0/core/src/main/...
      .../java/org/apache/calcite/plan/volcano/VolcanoCost.java

The only differences between this file and the Calcite file are:
1) All VolcanoCost strings have been changed to ImpalaCost
2) The package name is an Impala package.

This will make it easier to show the changes made for the Impala cost
model change in IMPALA-14101.

Change-Id: I864e20fb63c0ae4f2f88016128d2a68f39e17dfb
Reviewed-on: http://gerrit.cloudera.org:8080/23310
Reviewed-by: Riza Suminto <riza.suminto@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>