IMPALA-14409: Calcite planner: support statement_expression_limit
Added a counter in the place where the Calcite planner handles
expressions.  A RexNode in Calcite is similar to an Expr object
in Impala.

The test for this exists in
test_exprs.py::TestExprLimits::test_max_statement_size and will be
run once all jenkins tests are run through Calcite.

Change-Id: I9210a9a86eb98b5923295b846cb29f10acf459ce
Reviewed-on: http://gerrit.cloudera.org:8080/23765
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>