IMPALA-14488: Calcite planner: Support fallback to Original planner
Introduced the query options PLANNER and FALLBACK_PLANNER which are
of type TPlannerType

The 3 options for TPlannerType are:

ORIGINAL (default) : Use the Original planner for queries.
CALCITE            : Use the Calcite planner for queries.
NONE               : Convenient option for setting the FALLBACK_PLANNER.
                     This should never be used for the PLANNER option.

The Calcite planner currently does not support DDL and other various
queries (e.g. complex types). If a known unsupported query is detected,
the original planner will be used regardless of what is chosen for
FALLBACK_PLANNER.

The fallback planner will allow for easier transition for users who
want to use the new Calcite planner without having to worry about
backward compatibility issues that the compiler doesn't support.
One example that is currently not supported (in runtime_planners.test)
is the "ignore nulls" phrase which needs to be inside the parens for
Impala, but outside the parens for the Calcite parser.

At the time of this commit, there are also various compiler exceptions
thrown for Calcite which are not yet supported. The PLANNER=CALCITE
and FALLBACK_PLANNER=ORIGINAL options are set if the
--use_calcite_planner=true server option is specified.  This will allow
easier transitioning to the Calcite planner because the user will be
assured that even if there is a compilation failure in Calcite,
their existing query will still run with the original planner.

Change-Id: Id1fdc5ef92fff84e89af0e19c4246cc15e2ea823
Reviewed-on: http://gerrit.cloudera.org:8080/23900
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Michael Smith <michael.smith@cloudera.com>