IMPALA-14674: Implement connect_timeout_ms for HS2-HTTP
Implements connect_timeout_ms support for HS2-HTTP. This only applies
the timeout while establishing the connection, so later requests do not
timeout. This is preferrable to configuring http_socket_timeout_s, which
can result in timing out operations that just take awhile to execute.

Removed ImpalaHttpClient.setTimeout as it's unused and not part of the
specification for TTransportBase.

Testing:
- updates tests now that HS2-HTTP supports connect_timeout_ms
- test_impala_shell_timeout for HS2-HTTP without http_socket_timeout_s
- marks test_http_socket_timeout to run serially because it relies on
  short timeouts; inconsistent behavior can leave a dangling session

Change-Id: I9012066fb0d16497f309532021d7b323404b9fb2
Reviewed-on: http://gerrit.cloudera.org:8080/23499
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Tested-by: Michael Smith <michael.smith@cloudera.com>